bot/requirements.txt
pyodbc==4.0.30
SQLAlchemy==1.3.16
beautifulsoup4==4.9.1
requests==2.22.0
pandas==1.0.1
sklearn==0.0
python-telegram-bot==12.8